Hermaphrodites, and beginning of UI
Hermaphrodites have been added. Stats for hermaphrodites have been added. A state bug that didn't provide the option of returning to the "Wander" state if reproduction could occur has bee fixed in all reproductive types. UI for settings has begun implementation. IMPORTANT: The UI settings need to be changed to submit all settings when the "Run Simulation" button is pressed, and not in real time. This allow for the menu performance to not only be higher, but it will also allow the settings to not only occur when a change has occurred and instead set themselves to the default values provided.