Rework of Settings Selection
GeneralManager.cs has now been deleted, and all of it's contents have been put in MenuManager.cs . This allows for activating selected things to only happen once, when the "Run Simulation" button is pressed. This means things don't have to be set inactive outside of Start(). All SetActive() calls are now called on an array, and the index that is called is given from a specific dropdown. This entire rework will make adding new settings very easy, and scalable.