Predator Catch Rates
Predators have a non-zero chance of failing to catch prey. When predators collide with prey first they check to see if they are bigger then the prey. If the predator is bigger then it will calculate a size difference and roll a random float from 0f to that size difference. If that random float is greater than half of the size difference the prey will be caught and eaten.