Predator Sensory, Pause Bug. Fixed
Predators don't target prey to chase unless their size is bigger than or equal to the sensed prey. The bug where creatures and predators could die from lack of energy when the simulation was paused has been fixed. The energy loss is not multiplied by Time.timeScale to ensure that when the game is paused no energy is lost.